made few changes in  authen part
